Round 2 of fluid vecs
Waaay fewer classes & cases as before, also dropping ref-cnts.
Thin CVec compression POJO's over Value mem bufs.
frozenType is now a short everywhere

+++ src/main/java/water/DTask.java
@@ -49,5 +49,5 @@ public abstract class DTask<T> extends H2OCountedCompleter implements Freezable 
   @Override public AutoBuffer write(AutoBuffer bb) { throw barf(); }
   @Override public <F extends Freezable> F read(AutoBuffer bb) { throw barf(); }
   @Override public <F extends Freezable> F newInstance() { throw barf(); }
-  @Override public int frozenType() { throw barf(); }
+  @Override public short frozenType() { throw barf(); }
 }
